split route-options and expose `getRouteOptions`
