+++ src/runtime/index.ts
@@ -4,3 +4,4 @@ export * from './cache'
 export { useNitroApp } from './app'
 export * from './plugin'
 export * from './renderer'
+export { getRouteOptions } from './route-options'
